refactor: Introduce PRNotesPanelViewModel
